Add new features
- Refactor;
- Add elitism;
- Add parameters via CLI;
- Generate chromosome fitness evolution chart.

+++ internal/chromosome/external.go
@@ -6,12 +6,13 @@ var (
 	functionModel  fitness.FunctionModel
 	domainMin      float64
 	domainMax      float64
-	numberBits     int64
-	numberBitsEach int64
+	numberBits     int
+	numberBitsEach int
 	mutationRate   float64
 )
 
-func Initialize(funcModel fitness.FunctionModel, min, max float64, bits int64, mutation float64) {
+// Initialize set the genetic parameters
+func Initialize(funcModel fitness.FunctionModel, min, max float64, bits int, mutation float64) {
 	functionModel = funcModel
 	domainMin = min
 	domainMax = max
